pyproject.toml
[project]
name = "s_mapper_api"
version = "0.1.0"
description = "API for Smapper handheld device"
readme = "README.md"
license = "GPL-3.0"
requires-python = ">=3.10"
dependencies = [
    "coloredlogs>=15.0.1",
    "colorlog>=6.9.0",
    "fastapi[standard]>=0.115.2",
    "psutil>=7.0.0",
    "pybase62>=1.0.0",
    "pydantic-settings>=2.6.0",
    "pytest>=8.3.3",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
]
# dbus-python = "1.3.2"
# PyGObject = "3.50.0"
# lark = "^1.2.2"
# numpy = "^2.1.2"
